reusable tokenbox with errors
